requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
djangorestframework==3.15.2
git-filter-repo==2.45.0
idna==3.10
python-decouple==3.8
redis==5.2.0
requests==2.32.3
spotipy==2.24.0
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
PyJWT==2.10.0